moved ShaderControls to workaround namespace issues with Cinder-View
